FIX restore script

+++ lib/restore.sh
@@ -4,8 +4,8 @@ function restore () {
     [[ $# -eq "0" ]] && return
     [[ -d "${HOME}/${BACKUP_DIR}" ]] || create_backup
 
-    local TARGET=$(readlink -f "$1")
+    local TARGET=${1}
 
-    [[ -e "${HOME}/.dotfiles.backup/${TARGET//\//\!}" ]] | return
+    [[ -e "${HOME}/.dotfiles.backup/${TARGET//\//\!}" ]] || return
     command cp -rdnf ${HOME}/.dotfiles.backup/${TARGET//\//\!} ${TARGET}
 }
